seperate the controller and routes
